Ramailo_task1/requirements.txt
asgiref==3.6.0
autopep8==2.0.2
Django==4.1.7
django-jsonfield==1.4.1
django-rest-framework==0.1.0
djangorestframework==3.14.0
pycodestyle==2.10.0
pytz==2022.7.1
six==1.16.0
sqlparse==0.4.3
tomli==2.0.1
tzdata==2022.7